requirements.txt
Django==1.11.5
djangorestframework==3.6.4
markdown
django-filter==1.0.4
Pillow
